fix: :bug: fixed create, edit, delete in accordance to the user

+++ path/comments_path.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter,status, Query
+from fastapi import APIRouter,status, Query, HTTPException
 from dal import comments_dal
 from fastapi.params import Depends
 from database.database import get_db
@@ -6,6 +6,7 @@ from sqlalchemy.orm import Session
 from models import comments_models
 from security.security import get_current_user
 from models.users_models import User
+from models.comments_models import Comment
 
 
 
@@ -13,16 +14,19 @@ path=APIRouter()
 
 @path.post("/comments", status_code=status.HTTP_201_CREATED)
 def add_comment(request:comments_models.Comment , db: Session = Depends(get_db), current_user:User = Depends(get_current_user)):
-    return comments_dal.add_comment(request,db)
+    request.user_id = current_user.user_id
+    return comments_dal.add_comment(request,db,current_user)
 
 @path.put("/comments/{id}")
 def update_comment(id:int, request:comments_models.Comment, db: Session = Depends(get_db), current_user:User = Depends(get_current_user)):
-    comment.user_id = current_user.user_id
+    request.user_id = current_user.user_id
     return comments_dal.update_comment(id, request,db)
 
 @path.delete("/comments/{id}")
 def delete_comment(id:int, db: Session = Depends(get_db), current_user:User = Depends(get_current_user)):
-    comment.user_id = current_user.user_id
+    is_admin = current_user.user_type == 'administrateur'
+    if not is_admin:
+        raise HTTPException(status_code=403, detail="You do not have the right to delete this comment")
     return comments_dal.delete_comment(id,db)
 
 @path.get("/comments")
